Always upgrade elasticsearch plugin
We ensure the elasticsearch plugin for collectd is installed through
a git repo link. This means that, if it is already installed, we dont
get any updates. This change makes the default to install any upgrade
to the collectd elasticsearch plugin. This makes sense as we already
only specify master by default as the installation source.